+++ react/Reviews.js
@@ -13,6 +13,7 @@ import queryRatingSummary from './graphql/queries/queryRatingSummary.gql'
 import voteReviewQuery from './graphql/mutations/voteReview.gql'
 import { withApollo } from 'react-apollo'
 import { FormattedMessage, useIntl, defineMessages } from 'react-intl'
+import { SliderLayout } from 'vtex.slider-layout'
 
 import {
   IconSuccess,
@@ -24,6 +25,18 @@ import {
 import useFeedless from './modules/useFeedless'
 
 const IMAGES_URI_PREFIX = '//images.powerreviews.com'
+const ITEMS_PER_PAGE = {
+  desktop: 3,
+  tablet: 2,
+  phone: 2,
+}
+const INFINITE = true
+const NAVIGATION_STEP = 'page'
+const SHOW_NAVIGATION_ARROWS = 'desktopOnly'
+const SHOW_PAGINATION_DOTS = 'always'
+const USE_PAGINATION = true
+const FULL_WIDTH = false
+const ARROW_SIZE = 25
 
 const CSS_HANDLES = [
   'powerReviewsWrapper',
@@ -247,7 +260,7 @@ const Reviews = props => {
 
   const [state, dispatch] = useReducer(reducer, initialState)
   const { filter, selected, page, count, histogram, average } = state
-
+  const { reviewMediaAsSlider } = props
   const config = props.appSettings ? props.appSettings : {}
 
   useFeedless(config)
@@ -383,9 +396,23 @@ const Reviews = props => {
     )
   }
 
+  const renderImages = (review, widthClassName) =>
+    review.media.map((item, i) => {
+      return (
+        <img
+          alt=""
+          className={`${widthClassName} db mb5`}
+          key={i}
+          src={item.uri}
+        />
+      )
+    })
+
   const formattedOptions = options(formatMessage)
 
   const formattedFilters = filters(formatMessage)
+
+  const reversedHistogram = [...histogram].reverse()
 
   return (
     <div
@@ -404,26 +431,68 @@ const Reviews = props => {
         >
           {average}
         </span>
+        <span
+          className={`${handles.powerReviewsRatingSummaryAverage} review__rating--average ml3 c-muted-2`}
+        >
+          (
+          <FormattedMessage
+            id="store/power-reviews.numberOfReviews"
+            values={{ number: state.count }}
+          />
+          )
+        </span>
       </div>
       {state.reviews.length > 0 && (
         <div className={`${handles.powerReviewsHistogram} review__histogram`}>
           <ul className="bg-muted-5 pa7 list">
             {state.percentage.map((percentage, i) => {
               return (
-                <li key={i} className="mv3">
-                  <span className="dib w-10 v-mid">
-                    <FormattedMessage
-                      id="store/power-reviews.stars"
-                      values={{ stars: 5 - i }}
-                    />
-                  </span>
-                  <div className="review__histogram--bar bg-white dib h2 w-90 v-mid">
-                    <div
-                      className="review__histogram--bar-value h2 bg-yellow"
-                      style={{ width: percentage }}
-                    ></div>
-                  </div>
-                </li>
+                <div key={i}>
+                  <li key={i} className="mv3 flex-s items-center-s">
+                    <span
+                      className="dib v-mid mr3-s flex items-end"
+                      style={{ width: '5%' }}
+                    >
+                      <FormattedMessage
+                        id="store/power-reviews.stars"
+                        values={{ stars: 5 - i }}
+                      />{' '}
+                      <svg
+                        className="ml2 dib-ns dn"
+                        style={{
+                          verticalAlign: 'text-bottom',
+                        }}
+                        key={i}
+                        xmlns="http://www.w3.org/2000/svg"
+                        width="20"
+                        height="20"
+                        fill="#fc0"
+                        viewBox="0 0 14.737 14"
+                      >
+                        <path
+                          d="M7.369,11.251,11.923,14,10.714,8.82l4.023-3.485-5.3-.449L7.369,0,5.3,4.885,0,5.335,4.023,8.82,2.815,14Z"
+                          transform="translate(0)"
+                        />
+                      </svg>
+                    </span>
+                    <div className="review__histogram--bar bg-white dib h2 w-90 v-mid br4">
+                      <div
+                        className="review__histogram--bar-value h2 bg-yellow br4"
+                        style={{
+                          width: percentage,
+                          borderTopRightRadius: 0,
+                          borderBottomRightRadius: 0,
+                        }}
+                      />
+                    </div>
+                    <span
+                      className="dib w-10 v-mid ttl tc c-muted-2 w-0-s ml3-s"
+                      style={{ width: '5%' }}
+                    >
+                      ({reversedHistogram[i]})
+                    </span>
+                  </li>
+                </div>
               )
             })}
           </ul>
@@ -634,16 +703,23 @@ const Reviews = props => {
 
                 {review.media.length ? (
                   <div className="review__comment-images mt6">
-                    {review.media.map((item, i) => {
-                      return (
-                        <img
-                          alt=""
-                          className="w-20 db mb5"
-                          key={i}
-                          src={item.uri}
-                        />
-                      )
-                    })}
+                    {reviewMediaAsSlider && review.media.length > 3 ? (
+                      <SliderLayout
+                        totalItems={review.media?.items?.length}
+                        infinite={INFINITE}
+                        itemsPerPage={ITEMS_PER_PAGE}
+                        navigationStep={NAVIGATION_STEP}
+                        showNavigationArrows={SHOW_NAVIGATION_ARROWS}
+                        showPaginationDots={SHOW_PAGINATION_DOTS}
+                        usePagination={USE_PAGINATION}
+                        fullWidth={FULL_WIDTH}
+                        arrowSize={ARROW_SIZE}
+                      >
+                        {renderImages(review, 'w-90')}
+                      </SliderLayout>
+                    ) : (
+                      renderImages(review, 'w-20')
+                    )}
                   </div>
                 ) : null}
               </div>